+++ core/components/_helpers/automation-attribute.js
@@ -1,3 +1,3 @@
-export default (defaultKey) => ({
+export default defaultKey => ({
   'data-cosmos-key': defaultKey
-});
+})