+++ core/engine/configs/webpack.production.config.js
@@ -1,10 +1,13 @@
 const webpack = require('webpack')
 const path = require('path')
+const HtmlWebpackPlugin = require('html-webpack-plugin')
+const CleanWebpackPlugin = require('clean-webpack-plugin')
 
 module.exports = {
-  mode: 'production',
+  mode: 'development',
   entry: './index.js',
   output: {
+    filename: 'main.[contenthash].js',
     path: path.resolve(process.cwd(), 'public')
   },
   module: {
@@ -25,5 +28,11 @@ module.exports = {
   node: {
     fs: 'empty'
   },
-  plugins: []
+  plugins: [
+    new CleanWebpackPlugin(['*.js'], { root: path.resolve(process.cwd(), 'public') }),
+    new HtmlWebpackPlugin({
+      template: 'public/index.html',
+      inject: true
+    })
+  ]
 }