+++ changelog.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). Currently, this project is using an `0.x` versioning system to indicate its preview status. As such, it does _not_ currently adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html). Once we reach production readiness, we will follow semver beginning with our `1.0` release. In the meantime, please read this changelog for information on breaking changes!
 
+## 0.5.1 [September 3, 2018]
+
+### Added
+
+- Component: Danger Zone (experimental) [#847]
+
+### Fixed
+
+- Table: Add support for a Table that isn't sortable + improve documentation [#865]
+- Docs: Fix production version + implement cache busting [#876]
+
 ## 0.5.0 [August 30, 2018]
 
 ### Added